Change to how image origins are placed.
Modified the default behavior of the bitmap plots using imshow() to
place the image origin at the bottom left corner, matching the usual
behavior of transposing the field image before plotting, to avoid
accidentally flipping the image along y.

Also added a parameter to the matplotlib imshow() functions allowing the
user to set the origin location using the keywords recognized by
imshow(). The function to save an image with Pillow now has a note in
the documentation noting that the origin is at the top left and may
result in the image being flipped along the y-axis.